added my items

+++ src/Components/Pages/Home/AllInventory/AllInventory.js
@@ -1,29 +1,34 @@
-import React, { useState } from "react";
-import useInventory from "../../../../hooks/useInventory";
-import PageTitle from "../../Shared/PageTitle/PageTitle";
+import React, { useEffect, useState } from "react";
 import { toast } from "react-hot-toast";
+import PageTitle from "../../Shared/PageTitle/PageTitle";
 import ManageInventory from "../ManageInventory/ManageInventory";
 import "../ManageInventory/ManageInventory.css";
 
 const AllInventory = () => {
-  const [inventories, setInventories] = useInventory();
-  const [show, setShow] = useState(false);
+  const [inventories, setInventories] = useState([]);
+  useEffect(() => {
+    fetch("https://cars-warehouse.herokuapp.com/cars")
+      .then((res) => res.json())
+      .then((data) => setInventories(data));
+  }, []);
 
-  // Delete a inventory item
   const handleDeleteInventories = (id) => {
+    const confirm = window.confirm("Are you sure you want to delete");
+    if (!confirm) {
+      return;
+    }
     fetch(`https://cars-warehouse.herokuapp.com/cars/${id}`, {
       method: "DELETE",
     })
       .then((res) => res.json())
       .then((result) => {
-        toast.success("Inventory Deleted successfully!", {
-          duration: 4000,
-          position: "top-center",
-        });
+        if (result.deletedCount > 0) {
+          toast.success("Inventory Deleted Successfully");
+        }
         setInventories(inventories.filter((inventory) => inventory._id !== id));
       });
-    setShow(false);
   };
+
   return (
     <div className="container-fluid">
       <PageTitle title="Manage Items"></PageTitle>
@@ -37,8 +42,6 @@ const AllInventory = () => {
               <ManageInventory
                 inventories={inventories}
                 handleDeleteInventories={handleDeleteInventories}
-                show={show}
-                setShow={setShow}
               ></ManageInventory>
             }
           </div>
